reorganized csearch and its utils into a single module
